Update tests for client and pool
